Wall running class complete but needs testing
info: dropped the AddDynamic binds for OnJump and Landed in the BeginPlay function for WallRunningCharacter component due to jumped being non visible from the ACharacter class. Decided to modify my MovementSystemCharacter to manually notify the WallRunComponent when the player jumps. I had to make the properties above public to do so but hope to make a more elegant solution in the future.